requirements.txt
Flask
Flask-Login
mysqlclient
pillow
pytest
numpy